# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ClarkNav API",
        "description": "API documentation for ClarkNav",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/custom-routes": {
            "get": {
                "tags": [
                    "Custom Routes"
                ],
                "summary": "Get all custom routes",
                "operationId": "8ccf2908d8d5fdd44ae6c8f79d50cb93",
                "responses": {
                    "200": {
                        "description": "List of custom routes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "user_id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "origin": {
                                                "type": "string",
                                                "example": "Astro Park"
                                            },
                                            "destination": {
                                                "type": "string",
                                                "example": "Clark Airport Parking"
                                            },
                                            "route_name": {
                                                "type": "string",
                                                "example": "Route 1"
                                            },
                                            "fare": {
                                                "type": "number",
                                                "format": "float",
                                                "example": 50
                                            },
                                            "student_fare": {
                                                "type": "number",
                                                "format": "float",
                                                "example": 40
                                            },
                                            "duration": {
                                                "type": "string",
                                                "example": "1 hour"
                                            },
                                            "departure_time": {
                                                "type": "string",
                                                "example": "08:00 AM"
                                            },
                                            "arrival_time": {
                                                "type": "string",
                                                "example": "09:00 AM"
                                            },
                                            "departure_date": {
                                                "type": "string",
                                                "format": "date",
                                                "example": "2023-01-01"
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2023-01-01T00:00:00Z"
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2023-01-01T00:00:00Z"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Custom Routes"
                ],
                "summary": "Store a new custom route",
                "operationId": "4239612f0ae5a001515e4049ae8dac52",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "origin",
                                    "destination",
                                    "route_name",
                                    "fare",
                                    "duration",
                                    "departure_time",
                                    "arrival_time",
                                    "departure_date"
                                ],
                                "properties": {
                                    "origin": {
                                        "type": "string",
                                        "example": "Astro Park"
                                    },
                                    "destination": {
                                        "type": "string",
                                        "example": "Clark Airport Parking"
                                    },
                                    "route_name": {
                                        "type": "string",
                                        "example": "Route 1"
                                    },
                                    "fare": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 50
                                    },
                                    "student_fare": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 40
                                    },
                                    "duration": {
                                        "type": "string",
                                        "example": "1 hour"
                                    },
                                    "departure_time": {
                                        "type": "string",
                                        "example": "08:00 AM"
                                    },
                                    "arrival_time": {
                                        "type": "string",
                                        "example": "09:00 AM"
                                    },
                                    "departure_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2023-01-01"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Custom route created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "user_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "origin": {
                                            "type": "string",
                                            "example": "Astro Park"
                                        },
                                        "destination": {
                                            "type": "string",
                                            "example": "Clark Airport Parking"
                                        },
                                        "route_name": {
                                            "type": "string",
                                            "example": "Route 1"
                                        },
                                        "fare": {
                                            "type": "number",
                                            "format": "float",
                                            "example": 50
                                        },
                                        "student_fare": {
                                            "type": "number",
                                            "format": "float",
                                            "example": 40
                                        },
                                        "duration": {
                                            "type": "string",
                                            "example": "1 hour"
                                        },
                                        "departure_time": {
                                            "type": "string",
                                            "example": "08:00 AM"
                                        },
                                        "arrival_time": {
                                            "type": "string",
                                            "example": "09:00 AM"
                                        },
                                        "departure_date": {
                                            "type": "string",
                                            "format": "date",
                                            "example": "2023-01-01"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2023-01-01T00:00:00Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2023-01-01T00:00:00Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}